+++ app/scripts/parallel/corpus_listing.ts
@@ -3,63 +3,59 @@ import _ from "lodash"
 import settings from "@/settings"
 import { CorpusListing } from "@/corpus_listing"
 import { locationSearchGet } from "@/util"
+import { CorpusTransformed } from "@/settings/config-transformed.types"
+import { Attribute } from "@/settings/config.types"
+import { LangString } from "@/i18n/types"
 
 export class ParallelCorpusListing extends CorpusListing {
-    constructor(corpora) {
+    activeLangs: string[]
+
+    constructor(corpora: Record<string, CorpusTransformed>) {
         super(corpora)
 
-        const hash = window.location.hash.substr(2)
-        for (let item of hash.split("&")) {
-            var parts = item.split("=")
-            if (parts[0] == "parallel_corpora") {
-                this.setActiveLangs(parts[1].split(","))
-                break
-            }
-        }
+        // Cannot use Angular helpers (`locationSearchGet`) here, it's not initialized yet.
+        const activeLangs = new URLSearchParams(window.location.hash.slice(2)).get("parallel_corpora")
+        this.setActiveLangs(activeLangs.split(","))
     }
 
-    select(idArray) {
-        this.selected = []
-        $.each(idArray, (i, id) => {
-            const corp = this.struct[id]
-            this.selected = this.selected.concat(this.getLinked(corp, true, false))
-        })
-
-        this.selected = _.uniq(this.selected)
+    select(idArray: string[]): void {
+        this.selected = _.uniq(idArray.flatMap((id) => this.getLinked(this.struct[id])))
         this.updateAttributes()
     }
 
-    setActiveLangs(langlist) {
+    setActiveLangs(langlist: string[]): void {
         this.activeLangs = langlist
     }
 
-    getReduceLang() {
+    getReduceLang(): string {
         return this.activeLangs[0]
     }
 
-    getCurrentAttributes(lang) {
+    getCurrentAttributes(lang?: string): Record<string, Attribute> {
         if (_.isEmpty(lang)) {
             lang = settings.corpusListing.getReduceLang()
         }
 
         const corpora = _.filter(this.selected, (item) => item.lang === lang)
-        const struct = _.reduce(corpora, (a, b) => $.extend({}, a.attributes, b.attributes), {})
-        return struct
+        return corpora.reduce((attrs, corpus) => ({ ...attrs, ...corpus.attributes }), {} as Record<string, Attribute>)
     }
 
-    getStructAttrs(lang) {
+    getStructAttrs(lang?: string): Record<string, Attribute> {
         if (_.isEmpty(lang)) {
             lang = settings.corpusListing.getReduceLang()
         }
 
         const corpora = _.filter(this.selected, (item) => item.lang === lang)
-        const struct = _.reduce(corpora, (a, b) => $.extend({}, a["struct_attributes"], b["struct_attributes"]), {})
-        $.each(struct, (key, val) => (val["is_struct_attr"] = true))
+        const struct = corpora.reduce(
+            (attrs, corpus) => ({ ...attrs, ...corpus.struct_attributes }),
+            {} as Record<string, Attribute>
+        )
+        Object.values(struct).forEach((attr) => (attr.is_struct_attr = true))
 
         return struct
     }
 
-    getStructAttrsIntersection(lang) {
+    getStructAttrsIntersection(lang: string): Record<string, Attribute> {
         const corpora = _.filter(this.selected, (item) => item.lang === lang)
         const attrs = _.map(corpora, function (corpus) {
             for (let key in corpus["struct_attributes"]) {
@@ -72,45 +68,20 @@ export class ParallelCorpusListing extends CorpusListing {
         return this._mapping_intersection(attrs)
     }
 
-    getLinked(corp, andSelf, only_selected) {
-        if (andSelf == null) {
-            andSelf = false
-        }
-        if (only_selected == null) {
-            only_selected = true
-        }
-        const target = only_selected ? this.selected : this.struct
-        let output = _.filter(target, (item) => (corp["linked_to"] || []).includes(item.id))
-        if (andSelf) {
-            output = [corp].concat(output)
-        }
-        return output
+    getLinked(corp: CorpusTransformed, only_selected?: boolean) {
+        const target = only_selected ? this.selected : Object.values(this.struct)
+        let output: CorpusTransformed[] = target.filter((item) => (corp["linked_to"] || []).includes(item.id))
+        return [corp].concat(output)
     }
 
-    getEnabledByLang(lang, andSelf, flatten) {
-        if (andSelf == null) {
-            andSelf = false
-        }
-        if (flatten == null) {
-            flatten = true
-        }
+    getEnabledByLang(lang: string): CorpusTransformed[][] {
         const corps = _.filter(this.selected, (item) => item["lang"] === lang)
-        const output = _(corps)
-            .map((item) => {
-                return this.getLinked(item, andSelf)
-            })
-            .value()
-
-        if (flatten) {
-            return _.flatten(output)
-        } else {
-            return output
-        }
+        return corps.map((item) => this.getLinked(item, true))
     }
 
-    getLinksFromLangs(activeLangs) {
+    getLinksFromLangs(activeLangs: string[]): CorpusTransformed[][] {
         if (activeLangs.length === 1) {
-            return this.getEnabledByLang(activeLangs[0], true, false)
+            return this.getEnabledByLang(activeLangs[0])
         }
         // get the languages that are enabled given a list of active languages
         const main = _.filter(this.selected, (corp) => corp.lang === activeLangs[0])
@@ -131,7 +102,7 @@ export class ParallelCorpusListing extends CorpusListing {
         return output
     }
 
-    getAttributeQuery(attr) {
+    getAttributeQuery(attr: "context" | "within"): string {
         // gets the within and context queries
 
         const struct = this.getLinksFromLangs(this.activeLangs)
@@ -143,12 +114,7 @@ export class ParallelCorpusListing extends CorpusListing {
             const other = corps.slice(1)
 
             const pair = _.map(other, function (corp) {
-                let a
-                if (mainIsPivot) {
-                    a = _.keys(corp[attr])[0]
-                } else {
-                    a = _.keys(corps[0][attr])[0]
-                }
+                const a = mainIsPivot ? _.keys(corp[attr])[0] : _.keys(corps[0][attr])[0]
                 return mainId + "|" + corp.id.toUpperCase() + ":" + a
             })
             return output.push(pair)
@@ -157,17 +123,17 @@ export class ParallelCorpusListing extends CorpusListing {
         return output.join(",")
     }
 
-    getContextQueryString() {
+    getContextQueryString(): string {
         return this.getAttributeQuery("context")
     }
 
-    getWithinParameters() {
+    getWithinParameters(): { default_within: string; within: string } {
         const defaultWithin = locationSearchGet("within") || _.keys(settings["default_within"])[0]
         const within = this.getAttributeQuery("within")
         return { default_within: defaultWithin, within }
     }
 
-    stringifySelected(onlyMain) {
+    stringifySelected(onlyMain?: boolean): string {
         let struct = this.getLinksFromLangs(this.activeLangs)
         if (onlyMain) {
             struct = _.map(struct, (pair) => {
@@ -182,7 +148,6 @@ export class ParallelCorpusListing extends CorpusListing {
         }
 
         const output = []
-        // $.each(struct, function(i, item) {
         for (let i = 0; i < struct.length; i++) {
             const item = struct[i]
             var main = item[0]
@@ -194,11 +159,11 @@ export class ParallelCorpusListing extends CorpusListing {
         return output.join(",")
     }
 
-    get(corpusID) {
+    get(corpusID: string): CorpusTransformed {
         return this.struct[corpusID.split("|")[1]]
     }
 
-    getTitleObj(corpusID) {
+    getTitleObj(corpusID: string): LangString {
         return this.get(corpusID).title
     }
 }